Refactored garbageCollection.ts code into multiple files
