+++ packages/runtime/container-runtime/src/gc/gcHelpers.ts
@@ -4,6 +4,10 @@
  */
 
 import { ITelemetryGenericEvent } from "@fluidframework/common-definitions";
+import {
+	IGarbageCollectionNodeData,
+	IGarbageCollectionState,
+} from "@fluidframework/runtime-definitions";
 import { packagePathToTelemetryProperty } from "@fluidframework/runtime-utils";
 import { MonitoringContext } from "@fluidframework/telemetry-utils";
 import {
@@ -11,7 +15,7 @@ import {
 	runSweepKey,
 	throwOnTombstoneLoadKey,
 	throwOnTombstoneUsageKey,
-} from "./garbageCollectionConstants";
+} from "./gcDefinitions";
 
 /**
  * Consolidates info / logic for logging when we encounter unexpected usage of GC'd objects. For example, when a
@@ -58,4 +62,15 @@ export function shouldAllowGcTombstoneEnforcement(
 		return true;
 	}
 	return persistedGeneration === currentGeneration;
+}
+
+export function generateSortedGCState(gcState: IGarbageCollectionState): IGarbageCollectionState {
+	const sortableArray: [string, IGarbageCollectionNodeData][] = Object.entries(gcState.gcNodes);
+	sortableArray.sort(([a], [b]) => a.localeCompare(b));
+	const sortedGCState: IGarbageCollectionState = { gcNodes: {} };
+	for (const [nodeId, nodeData] of sortableArray) {
+		nodeData.outboundRoutes.sort();
+		sortedGCState.gcNodes[nodeId] = nodeData;
+	}
+	return sortedGCState;
 }
